商城-卡密-活动三大块，以及一些权限的简单鉴定。 (需加：redis缓存商城数据，token+拦截器试试)

+++ dbo-backend/src/main/java/com/cc/controller/dbochar/DBOCharController.java
@@ -16,7 +16,7 @@ public class DBOCharController {
     private IDBOCharService dboCharService;
 
     @GetMapping("/list")
-    public Result getDBOCharList(Integer accountID){
+    public Result getDBOCharList(Long accountID){
         return dboCharService.getDBOCharList(accountID);
     }
 }
