+++ dbo-backend/src/main/java/com/cc/entity/User.java
@@ -4,16 +4,16 @@ import lombok.Data;
 
 @Data
 public class User {
-    private Integer accountID;
+    private Long accountID;
     private String username;
     private String accStatus;
     private String email;
     // 余额
-    private Integer mallPoints;
+    private Long mallPoints;
     // 现金扭蛋币
-    private Integer waguCoins;
+    private Long waguCoins;
     // 活动扭蛋币
-    private Integer eventCoins;
+    private Long eventCoins;
     // 是否为管理员
     private Integer admin;
     // 用户登录凭证
