+++ dbo-backend/src/main/java/com/cc/service/dbochar/IDBOCharService.java
@@ -3,5 +3,5 @@ package com.cc.service.dbochar;
 import com.cc.vo.Result;
 
 public interface IDBOCharService {
-    Result getDBOCharList(Integer accountID);
+    Result getDBOCharList(Long accountID);
 }
